app/requirements.txt
amqp==2.5.2
billiard==3.6.3.0
certifi==2020.4.5.1
Django==2.2.7
django-compat==1.0.15
django-cors-headers==3.1.1
django-extensions==2.2.6
djangorestframework==3.10.3
factory-boy==2.12.0
Faker==4.0.1
gunicorn==19.9.0
kombu==4.6.8
mysqlclient==1.4.6
PyJWT==1.7.1
python-dateutil==2.8.1
python-dotenv==0.10.3
pytz==2019.3
sentry-sdk==0.14.3
six==1.14.0
sqlparse==0.3.0
text-unidecode==1.3
urllib3==1.25.9
vine==1.3.0
django-crontab==0.7.1
